+++ src/action_request.cc
@@ -52,7 +52,7 @@ ActionRequest::ActionRequest(std::shared_ptr<UpnpXMLBuilder> xmlBuilder, std::sh
 {
     auto ctrlPtIPAddr = std::make_shared<GrbNet>(UpnpActionRequest_get_CtrlPtIPAddr(upnpRequest));
     std::string userAgent = UpnpActionRequest_get_Os_cstr(upnpRequest);
-    quirks = std::make_unique<Quirks>(std::move(xmlBuilder), clients, ctrlPtIPAddr, userAgent);
+    quirks = std::make_shared<Quirks>(std::move(xmlBuilder), clients, ctrlPtIPAddr, userAgent);
 }
 
 std::string ActionRequest::getActionName() const
@@ -70,7 +70,7 @@ std::string ActionRequest::getServiceID() const
     return serviceID;
 }
 
-const std::unique_ptr<Quirks>& ActionRequest::getQuirks() const
+const std::shared_ptr<Quirks>& ActionRequest::getQuirks() const
 {
     return quirks;
 }