+++ src/request_handler/file_request_handler.cc
@@ -175,7 +175,7 @@ bool FileRequestHandler::getInfo(const char* filename, UpnpFileInfo* info)
 
         // Generate DNLA Headers
         auto mappings = config->getDictionaryOption(ConfigVal::IMPORT_MAPPINGS_MIMETYPE_TO_CONTENTTYPE_LIST);
-        std::string dlnaContentHeader = xmlBuilder->getDLNAContentHeader(getValueOrDefault(mappings, mimeType), resource);
+        std::string dlnaContentHeader = xmlBuilder->getDLNAContentHeader(getValueOrDefault(mappings, mimeType), resource, quirks);
         if (!dlnaContentHeader.empty()) {
             headers.addHeader(UPNP_DLNA_CONTENT_FEATURES_HEADER, dlnaContentHeader);
         }