+++ src/upnp/xml_builder.h
@@ -80,7 +80,7 @@ public:
     ///
     /// This function looks at the object, and renders the DIDL-Lite representation of it -
     /// either a container or an item
-    void renderObject(const std::shared_ptr<CdsObject>& obj, std::size_t stringLimit, pugi::xml_node& parent, const std::unique_ptr<Quirks>& quirks = nullptr) const;
+    void renderObject(const std::shared_ptr<CdsObject>& obj, std::size_t stringLimit, pugi::xml_node& parent, const std::shared_ptr<Quirks>& quirks = nullptr) const;
 
     /// \brief Renders XML for the event property set.
     /// \return pugi::xml_document representing the newly created XML.
@@ -105,7 +105,7 @@ public:
 
     std::string renderResourceURL(const CdsObject& item, const CdsResource& res, const std::map<std::string, std::string>& mimeMappings, const std::string& clientGroup = "") const;
 
-    void addResources(const std::shared_ptr<CdsItem>& item, pugi::xml_node& parent, const std::unique_ptr<Quirks>& quirks) const;
+    void addResources(const std::shared_ptr<CdsItem>& item, pugi::xml_node& parent, const std::shared_ptr<Quirks>& quirks) const;
 
     /// \brief build path for first resource from item
     /// depending on the item type it returns the url to the media
@@ -114,7 +114,7 @@ public:
     /// \brief convert xml tree to string
     static std::string printXml(const pugi::xml_node& entry, const char* indent = PUGIXML_TEXT("\t"), int flags = pugi::format_default);
 
-    std::string getDLNAContentHeader(const std::string& contentType, const std::shared_ptr<CdsResource>& res) const;
+    std::string getDLNAContentHeader(const std::string& contentType, const std::shared_ptr<CdsResource>& res, const std::shared_ptr<Quirks>& quirks) const;
     std::string getDLNATransferHeader(const std::string& mimeType) const;
 
 protected:
@@ -131,7 +131,7 @@ protected:
     std::map<std::string, std::string> transferMappings;
 
     std::deque<std::shared_ptr<CdsResource>> getOrderedResources(const CdsObject& object) const;
-    std::pair<bool, int> insertTempTranscodingResource(const std::shared_ptr<CdsItem>& item, const std::unique_ptr<Quirks>& quirks, std::deque<std::shared_ptr<CdsResource>>& orderedResources, bool skipURL) const;
+    std::pair<bool, int> insertTempTranscodingResource(const std::shared_ptr<CdsItem>& item, const std::shared_ptr<Quirks>& quirks, std::deque<std::shared_ptr<CdsResource>>& orderedResources, bool skipURL) const;
 
     std::string renderExtension(const std::string& contentType, const fs::path& location, const std::string& language) const;
     void addField(pugi::xml_node& entry, const std::string& key, const std::string& val) const;
@@ -142,10 +142,10 @@ protected:
         const std::map<std::string, std::string>& auxData,
         ConfigVal itemProps,
         ConfigVal nsProp) const;
-    std::string findDlnaProfile(const CdsResource& res, const std::string& contentType) const;
-    std::string dlnaProfileString(const CdsResource& res, const std::string& contentType, bool formatted = true) const;
+    std::string findDlnaProfile(const CdsResource& res, const std::string& contentType, const std::shared_ptr<Quirks>& quirks) const;
+    std::string dlnaProfileString(const CdsResource& res, const std::string& contentType, const std::shared_ptr<Quirks>& quirks, bool formatted = true) const;
 
-    std::string buildProtocolInfo(CdsResource& res, const std::map<std::string, std::string>& mimeMappings) const;
+    std::string buildProtocolInfo(CdsResource& res, const std::map<std::string, std::string>& mimeMappings, const std::shared_ptr<Quirks>& quirks) const;
     std::string getMimeType(const CdsResource& resource, const std::map<std::string, std::string>& mimeMappings) const;
 };
 #endif // __UPNP_XML_H__